+++ frontend/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Created ProductCard and ProductCarousel, using the Swiper lib
+- Added custom colors and fonts to Tailwind in index.css
 - Added page title and favicon in "index.html"
 - Installed Tailwind following: https://tailwindcss.com/docs/installation/using-vite
 - Created boilerplate React v19.2.6 project using Vite v8.0.12 with Node LTS v24.16.0